chore: move enums and second level nested classes into options package
