+++ playwright/src/main/java/com/microsoft/playwright/Keyboard.java
@@ -16,6 +16,7 @@
 
 package com.microsoft.playwright;
 
+import com.microsoft.playwright.options.*;
 import java.util.*;
 
 /**
