+++ playwright/src/main/java/com/microsoft/playwright/impl/BrowserContextImpl.java
@@ -20,6 +20,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.microsoft.playwright.*;
+import com.microsoft.playwright.options.*;
 
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
