+++ playwright/src/main/java/com/microsoft/playwright/impl/Serialization.java
@@ -20,6 +20,7 @@ import com.google.gson.*;
 import com.google.gson.reflect.TypeToken;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
+import com.microsoft.playwright.options.*;
 import com.microsoft.playwright.*;
 
 import java.io.ByteArrayOutputStream;
