+++ playwright/src/main/java/com/microsoft/playwright/options/ColorScheme.java
@@ -14,13 +14,10 @@
  * limitations under the License.
  */
 
-package com.microsoft.playwright;
+package com.microsoft.playwright.options;
 
-public enum ElementState {
-  VISIBLE,
-  HIDDEN,
-  STABLE,
-  ENABLED,
-  DISABLED,
-  EDITABLE
+public enum ColorScheme {
+  LIGHT,
+  DARK,
+  NO_PREFERENCE
 }
