+++ playwright/src/main/java/com/microsoft/playwright/options/WaitUntilState.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.microsoft.playwright;
+package com.microsoft.playwright.options;
 
-public enum MouseButton {
-  LEFT,
-  RIGHT,
-  MIDDLE
+public enum WaitUntilState {
+  LOAD,
+  DOMCONTENTLOADED,
+  NETWORKIDLE
 }
