+++ playwright/src/test/java/com/microsoft/playwright/TestPageExposeFunction.java
@@ -21,7 +21,7 @@ import org.junit.jupiter.api.Test;
 import java.util.Map;
 
 import static com.microsoft.playwright.Utils.mapOf;
-import static com.microsoft.playwright.WaitUntilState.LOAD;
+import static com.microsoft.playwright.options.WaitUntilState.LOAD;
 import static java.util.Arrays.asList;
 import static org.junit.jupiter.api.Assertions.*;
